ipmock: rename pyodide to ipmock
and make it a generic IPRoute mocking class
