+++ setup.cfg
@@ -52,7 +52,7 @@ pyroute2 =
     DL = pyroute2.devlink:DL
     IPRoute = pyroute2.iproute:IPRoute
     IPBatch = pyroute2.iproute:IPBatch
-    IPMock = pyroute2.iproute.pyodide:IPRoute
+    IPMock = pyroute2.iproute.ipmock:IPRoute
     RawIPRoute = pyroute2.iproute:RawIPRoute
     ChaoticIPRoute = pyroute2.iproute:ChaoticIPRoute
     NetlinkError = pyroute2.netlink.exceptions:NetlinkError
