+++ setup.minimal.cfg
@@ -76,6 +76,7 @@ pyroute2 =
     IW = pyroute2.iwutil:IW
     IPRoute = pyroute2.iproute:IPRoute
     IPBatch = pyroute2.iproute:IPBatch
+    IPMock = pyroute2.iproute.ipmock:IPRoute
     RawIPRoute = pyroute2.iproute:RawIPRoute
     ChaoticIPRoute = pyroute2.iproute:ChaoticIPRoute
     NetlinkError = pyroute2.netlink.exceptions:NetlinkError
